Refactor PDF page rendering into separate PDFPage component
Extracted the PDF page rendering logic from PDFViewer.tsx into a new PDFPage.tsx component for improved modularity and maintainability. Updated PDFViewer to use the new PDFPage component and enhanced zoom controls to better preserve scroll position. Minor UI and code cleanups included.
